requirements.txt
pandas
numpy
embedchain
# camelot-py[base]
opencv-python
pdfplumber
requests
beautifulsoup4
selenium
webdriver-manager
langchain
pandas
faiss-cpu
transformers
sentence-transformers
lxml_html_clean
requests_html
pymupdf